ui push, come back after lunch and finish tying together the looks everything is working how it should
